changed hp percent from int to float

+++ GhoulKIng/Assets/Scripts/playerController.cs
@@ -286,7 +286,7 @@ public class playerController : MonoBehaviour, IDamageable
     }
     public void updatePlayerHP()
     {
-        int hpPercent = (HP / HPOrig * 100);
+        float hpPercent = ((float)HP / (float)HPOrig * 100);
         gameManager.instance.HPBar.fillAmount = (float)HP / (float)HPOrig;
         gameManager.instance.HPpercent.text = hpPercent.ToString("F0");
     }
